Move andor from SuperTypeExpression to ConstraintExpr
